Add new example to read and publish CAN data
This is an extension of CANReader.

This will read CAN data using the C++ CAN API, and populuate an instance
of the `ps_can_frame_msg` and publish it to the bus.

This was requested by a customer, and also allows for raw CAN data to be
viewable in Trace plugin.