Update the Pub/Sub code for the new tutorial
Changed the message type from event + diagnostic state, to platform
motion.

Seperated into two sepearate nodes.

Open to ideas if we want to make it more complex, such as adding the
ability for the publishing node to subscribe to it's own message, and
increase certain fields. This may be nices than pubilshing constant
values to the bus.