Add the final state of Ground Plane Detection (#30)
Studio needs to be tweaked with timeouts, since it doesn't persist the
ground plane points. It's possible that can be solved by playing with
this nodes timing.

Tweaked the point filtering to compensate for vehicle pitch. It was
common that the driver side ground plane points were filtered out since
the road had a curve. The car is driving near the apex, and the curve of
the road was commonly >0.25 meters.