Fix explicit settings of PSYNC_HOME
Prior to this commit, the CMakeLists.txt of the ParameterGetSet example
explicitly set PSYNC_HOME to /usr/local/polysync instead of checking for
an environment variable first, which caused errors if a user's machine
had PSYNC_HOME elsewhere. This commit first checks for an environment
variable before setting it explicitly.