Modularize and rename.
Prior code was named "Video*". This naming was inconsistent with
the message type and the name of the example. This commit moves
naming from Video* to ImageData. This commit also separates
functionality into three categories. PolySync Node, Image Processing,
and Image View. This separation makes represents the
responsibility of each element.