Maint/refactor (#61)
* Update example to use polysync::VideoDecoder.

Prior example was not functioning for MJPEG. This commit
adds the use of a polysync::VideoDecoder object to properly
decode incoming compressed data from MJPEG to RGB24.

* General cleanup.

Prior commit was allocating memory in main.cpp unnecessarily.
This commit switches to using the stackin main and moves
threading logic for VideoProcessor into it's class definition.
These changes reduce code side and improve readability.

* Modularize and rename.

Prior code was named "Video*". This naming was inconsistent with
the message type and the name of the example. This commit moves
naming from Video* to ImageData. This commit also separates
functionality into three categories. PolySync Node, Image Processing,
and Image View. This separation makes represents the
responsibility of each element.

* Fix #define wrapper names.