Update documentation for setting Qt environment.
Prior to this commit the a lack of documentation around the Qt image
viewer
left open the potential for a misconfigured Qt development environment.
This results in example programs crashing with Qt errors. After this
commit the documentation has been updated with steps require to
set Qt development environment required for building and running
this example program.