Update import/export example logic
Prior to this commit, LogSessionImport and LogSessionExport examples were reporting errors.
This exposed an existing bug in polysync-core-cpp-api. This commit adds some improvements
to the logic, including resource management for the _importer and _exporter member variables.
Following the bugfix (CORE-653) merge, this commit will enable proper functionality for
LogSessionImport and LogSessionExport examples.