Option to forgo writing to sdtout
Prior to this commit this example always wrote to stdout.
This meant the users who were interested in writing to a
specified file were unable to minimize the ammout of IO
work the tool had to do when not interested terminal reporting.

This commit represents the addition of a `-O` option to compliment
the existing `-o`. If `-O` is selected the example will
write to a file and not to stdout as well.

Logic was also added to report an error and activate a fatal
state if the example is unable to open the specified input
file rather that failing silently and maintaining the OK
state.